requirements.txt
transformers
torch
torchvision
torchaudio
pandas
sentencepiece
evaluate
sacrebleu
scikit-learn
tqdm
accelerate
bitsandbytes
pyyaml
tensorboard
